+++ mayan/apps/smart_settings/classes.py
@@ -8,10 +8,8 @@ import sys
 import yaml
 
 from django.conf import settings
+from django.utils.encoding import force_bytes, force_text
 from django.utils.functional import Promise
-from django.utils.encoding import (
-    force_bytes, force_text
-)
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.class_mixins import AppsModuleLoaderMixin